Fix circular import in Python bindings by renaming typing.py
