+++ runtime/bindings/python/iree/runtime/_binding.pyi
@@ -11,7 +11,7 @@ from typing import (
 )
 import asyncio
 
-from .typing import HalModuleBufferViewTraceCallback
+from ._typing import HalModuleBufferViewTraceCallback
 
 def create_hal_module(
     instance: VmInstance,
